+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example on how to create operations and use data preprocessing source to train models
   https://intel.github.io/dffml/examples/ice_cream.html
 - Operations for zip and tar file creation and extraction
+- Operations for file (de)compression
 ### Changed
 - Calls to hashlib now go through helper functions
 - Build docs using `dffml service dev docs`